requirements-venv.txt
# Minimal runtime deps for running Tensorflow/scripts/generate_tfrecord.py inside .venv
# Tested on Windows + Python 3.11
tensorflow==2.16.1
pandas
pillow
